backend/requirements.txt
Flask==2.3.3
python-dotenv==1.0.0
Flask-PyMongo==2.3.0
Flask-JWT-Extended==4.5.3
Flask-RESTX==1.3.0
pymongo==4.6.0
Flask-Limiter==3.8.0
Flask-CORS==6.0.1
redis==5.0.1